// src/Service/api/model/CheckoutModel.ts
export interface ICheckout {
    firstName: string;
    lastName: string;
    CompanyName: string;
    zipCode: string;
    country: string;
    street: string;
    town: string;
    province: string;
    addressSecond: string;
    email: string;
    additionalInformation: string;
    payment:string;
  }